+++ packages/capabilities/extraction/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./model";
+export * from "./ports";
+export * from "./normalise";
+export * from "./service";
+export * from "./manifest";