+++ packages/packs/jurisdiction-es-es/src/index.ts
@@ -3,4 +3,6 @@ export * from "./tax/adapter";
 export * from "./retention";
 export * from "./chain";
 export * from "./labels";
+export * from "./extraction/taxid";
+export * from "./extraction/profile";
 export * from "./manifest";